Archives Error 500 - Refactor and unify clear button component
Removed the old ClearButton and button-clear_ implementations, consolidating to a single ClearBtn component. Updated imports and usage in sidebar components to use the unified ClearBtn. Simplified the clear button logic and UI, removing redundant state and store interactions.
